requirements.txt
aiosql==9.0
aiosqlite==0.19.0
black==23.7.0
click==8.1.6
flake8==6.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pyaes==1.6.1
pyasn1==0.5.0
pycodestyle==2.11.0
pyflakes==3.1.0
python-dotenv==1.0.0
rsa==4.9
Telethon==1.29.2
tomli==2.0.1
